% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Data.R
\name{Extract.Data}
\alias{Extract.Data}
\title{Extract data from a TECAN measurement}
\usage{
Extract.Data(
  filepath,
  datagroups,
  datavalue.name = "Value",
  datagroup.leader = "Coordinate",
  na.warn = TRUE,
  na.rm = FALSE,
  .datagroups.prepostfix = TRUE,
  .debug = FALSE
)
}
\arguments{
\item{filepath}{Path to the input file}

\item{datagroups}{The user-defined data groups. They follow the pattern list(group1 = list(pattern1, pattern2), group2 = list(pattern1, pattern2))}

\item{datavalue.name}{The name of the value column.}

\item{datagroup.leader}{The column to match the groups against. This is recommended to leave unchanged.}

\item{na.warn}{Whether or not to show an error message when NAs are detected}

\item{na.rm}{Whether or not to remove NAs from the data. Only turn this on after verifying each instance of NAs}

\item{.datagroups.prepostfix}{This adds some sensible defaults to the regexp. Only ever touch this if you know what ^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$ means.}

\item{.debug}{Toggle debug mode. Does not do much for now.}
}
\value{
data.frame containing all measurements in long format.
}
\description{
This function loads an excel file that it assumes is in a TECAN generated
format. It then predicts what kind of data you generated and classifies each dataset based on
multiple heuristics.
}
