% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Layout.R
\name{Plot.Layout}
\alias{Plot.Layout}
\title{Preview the input group layout.}
\usage{
Plot.Layout(
  datagroups,
  datagroups.layout.Rows = LETTERS[1:8],
  datagroups.layout.Columns = 1:12,
  layoutGraph.scaleFactor = 0.6825,
  layoutGraph.dimensions = c(29.7, 21),
  .datagroups.prepostfix = TRUE
)
}
\arguments{
\item{datagroups}{The user-defined data groups. They follow the pattern list(group1 = list(pattern1, pattern2), group2 = list(pattern1, pattern2))}

\item{datagroups.layout.Rows}{The row values. By default for a 96-Well Plate this would be LETTERS[1:8] (A, B, C, D, E, F, G, H)}

\item{datagroups.layout.Columns}{The row values. By default for a 96-Well Plate this would be 1:12 (1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12)}

\item{layoutGraph.scaleFactor}{The scale factor for displaying the preview plot. I find a scale factor of 0.6825 to work for _me_. It might not work well on _your_ display.}

\item{layoutGraph.dimensions}{The dimensions (in cm) of the preview plot. By default A4 paper size in cm: c(29.7, 21.0),}

\item{.datagroups.prepostfix}{This adds some sensible defaults to the regexp. Only ever touch this if you understand the following pattern: ^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\\d]{8,}$}
}
\value{
ggplot2::plot
}
\description{
This function takes in datagroups and plots them to visually communicate the datagroups.
}
