% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Value.R
\name{Plot.Value}
\alias{Plot.Value}
\title{Preview the input group layout.}
\usage{
Plot.Value(
  data,
  data.preview.value = "Value",
  datagroups.layout.Rows = LETTERS[1:8],
  datagroups.layout.Columns = 1:12,
  data.preview.scaleFactor = 0.6825,
  data.preview.dimensions = c(29.7, 21),
  data.preview.showPreview = TRUE
)
}
\arguments{
\item{data}{the data.frame containing data output by one of the Extract.* functions}

\item{data.preview.value}{The value to compare in the plot. By default it is "Value"}

\item{datagroups.layout.Rows}{The row values. By default for a 96-Well Plate this would be LETTERS[1:8] (A, B, C, D, E, F, G, H)}

\item{datagroups.layout.Columns}{The row values. By default for a 96-Well Plate this would be 1:12 (1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12)}

\item{data.preview.scaleFactor}{The scale factor for displaying the preview plot. I find a scale factor of 0.6825 to work for _me_. It might not work well on _your_ display.}

\item{data.preview.dimensions}{The dimensions (in cm) of the preview plot. By default A4 paper size in cm: c(29.7, 21.0),}

\item{data.preview.showPreview}{Whether or not to show the preview.}
}
\value{
ggplot2::plot
}
\description{
This function takes in data and plots it in a spatially similar format as the Plot Layout.
}
